+++ CHANGELOG.md
@@ -284,3 +284,4 @@
 * Define `Generate` pattern
 * Remove `Component` typeclass
 * Change `Semimonoidal` type family definition
+* Change `Monoidal` type family definition